move RegimePurposeScopesConfig to core
